Add database connection
Previously, the web application did not connect to a database.  This
change adds postgresql database connection support.  The database to
be connected to is set using the $DATABASE_URL environment variable.
In addition, Flyway has been added to handle database schema
migrations.

Finally, the build system was moved from Gradle to Maven due to the
troubles in configuring AspectJ weaving.

[#54087197]