Update after_success to handle non-master
Previously, the Travis CI configuration of the project only build
master.  This was problematic as it meant that dev branches were not
built when pushed.  In addition, when pull requests came in, they were
built but failed because the secure environment variables were not set
when the build finished and it attempted to push.  This change updates
the configuration to build all branches, but to only push is the
branch is called master.

[#55027388]