Signatory Authentication
This change adds authentication for signatories.  Development of this
was a very enlightening experience.  It allowed me to see that there
was a bunch of code that I did not need and further identified code
that I did not understand.  The outcome of this is that there is less
complicated code in place.

A byproduct of this however was that it showed that I couldn't easily
have two different scopes for different kinds of users.  This was fine
as my research into statuses indicates that we couldn't use them (the
driving reason for an 'admin' scope).  If statuses change in the
future, I might have to re-attempt a second scope.

[#55221198]