Persist Signatory Data
This change adds support for persisting both individual and corporate
signatory data.  To do this, the data persistence layer was changed
from raw JDBC to JPA.  In addition, the repository implementations
were removed complete and replaced with Spring Data which dynamically
creates all implementations.  The final change was to simplify most of
the GitHub types such that they are not separate interfaces and
implementations.

[#55307532][#55307544]